First version of gopatterns.py
First version of gopatterns.py
Works with one, predefined pattern size, and also with preset min/max number of stones in the pattern.

Takes into account rotations, translation, color swap.

Attaches edges/corners when possible.

Does not match if there are stones outside of the pattern that are in contact with stones inside the pattern.